Refactor FileTypeSelector
